Add ArrayList DB and start add animal, Grazioso.
This commit adds the ArrayList to Driver.java to
serve as the animal DB, while also starting
the functionality to add a specific animal to the DB.
If-else statment executes and creates the appropriate
type based on the option selected.
The true method to add a full animal to the DB is still o/s.